requirements.txt
pillow
pyrogram